# server/api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Boards": {
      "get": {
        "tags": [
          "Boards"
        ],
        "operationId": "Boards_GetMyBoards",
        "parameters": [
          {
            "name": "includeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Boards"
        ],
        "operationId": "Boards_CreateBoard",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Board": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userId": {
            "type": "string",
            "format": "guid"
          },
          "numbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "repeating": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deleted": {
            "type": "boolean"
          },
          "gameId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "game": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Game"
              }
            ]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Game": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "week": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "winningNumbers": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deleted": {
            "type": "boolean"
          },
          "boards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "deleted": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "role": {
            "type": "string"
          },
          "salt": {
            "type": "string",
            "format": "guid"
          },
          "passwordHash": {
            "type": "string"
          },
          "boards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "Transaction": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "mobilepayReference": {
            "type": "string"
          },
          "approved": {
            "type": "boolean",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deleted": {
            "type": "boolean",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "format": "guid"
          },
          "user": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          }
        }
      },
      "CreateBoardDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userId",
          "gameId",
          "price",
          "repeating"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "gameId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "numbers": {
            "type": "array",
            "maxItems": 8,
            "minItems": 5,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "maximum": 2147483647.0,
            "minimum": 0.0
          },
          "repeating": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}